+++ mmpose/datasets/datasets/top_down/topdown_mpii_dataset.py
@@ -80,9 +80,9 @@ class TopDownMpiiDataset(TopDownBaseDataset):
             center = center - 1
 
             joints_3d = np.zeros((self.ann_info['num_joints'], 3),
-                                 dtype=np.float)
+                                 dtype=np.float32)
             joints_3d_visible = np.zeros((self.ann_info['num_joints'], 3),
-                                         dtype=np.float)
+                                         dtype=np.float32)
             if not self.test_mode:
                 joints = np.array(a['joints'])
                 joints_vis = np.array(a['joints_vis'])
@@ -194,7 +194,7 @@ class TopDownMpiiDataset(TopDownBaseDataset):
 
         # save
         rng = np.arange(0, 0.5 + 0.01, 0.01)
-        pckAll = np.zeros((len(rng), 16))
+        pckAll = np.zeros((len(rng), 16), dtype=np.float32)
 
         for r, threshold in enumerate(rng):
             less_than_threshold = (scaled_uv_err <= threshold) * jnt_visible
