+++ tests/test_pipelines/test_top_down_pipelines.py
@@ -82,8 +82,8 @@ def test_top_down_pipeline():
     ann = coco.anns[ann_ids[0]]
 
     num_joints = 17
-    joints_3d = np.zeros((num_joints, 3), dtype=np.float)
-    joints_3d_visible = np.zeros((num_joints, 3), dtype=np.float)
+    joints_3d = np.zeros((num_joints, 3), dtype=np.float32)
+    joints_3d_visible = np.zeros((num_joints, 3), dtype=np.float32)
     for ipt in range(num_joints):
         joints_3d[ipt, 0] = ann['keypoints'][ipt * 3 + 0]
         joints_3d[ipt, 1] = ann['keypoints'][ipt * 3 + 1]
